+++ CHANGELOG.md
@@ -17,6 +17,7 @@
  - switch the Event Dashboard button in Activity detail page when Guests Pending
  - add Subscription's Badge, using highest Privilege color,  on Calendar Weekly/Monthly Dots #386
  - update Subscripton's Badge, using hiest Privilege color, on Activity page #309 & #309
+ - fix #590 - wrong badges on activities by Theme when activity has event in past with pending subscriptions
 
 ## NEXT RELEASE
  - add pagination #585